squash! cmake: Check for availability of softfp libararies
Squash this into existing commit and continue the cleanup.

Feel free to adjust existing commit message with following text:

This commit extends existing `nrfxlib_calculate_lib_path()` function
to take a BASE_DIR.

The BASE_DIR allows the caller to specify a BASE_DIR from where the
lib path should be looked up.
When using a BASE_DIR the lib_path will be checked for existence before
the function returns. If the combined base directory and calculated
library path does not exists, the return variable `lib_path` will be set
to `${lib_path}-NOTFOUND` which means that the caller can use a simple
`if(<lib_path>)` construct to check the existence of the library path.

As addition, an optional `SOFT_FLOAT_FALLBACK` flag has been implemented
which allows a caller to request that a `soft-float` library path is
checked for existence if `softfp` ABI is selected but no `softfp`
library path was found.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>